pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-ai-assistant"
version = "0.1.0"
description = "Personal AI Assistant with multimodal storage, semantic search, and RAG architecture"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",

    # Vector DB
    "qdrant-client>=1.7.0",

    # Cache & Workers
    "redis>=5.0.0",
    "celery[redis]>=5.3.0",

    # AWS
    "boto3>=1.34.0",

    # AI / ML
    "anthropic>=0.40.0",
    "sentence-transformers>=3.0.0",
    "tiktoken>=0.6.0",

    # Document processing
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "Pillow>=10.2.0",

    # Image processing
    "ultralytics>=8.1.0",

    # Config
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.setuptools.packages.find]
include = ["app*"]

